Fixed some warnings, fixed compilation on a few utils
* Made all pmd2 projects compile.
* All pmd3 projects won't build. There seems to be something messing
with the build settings because it can't link the Poco libraries because
apparently the project is compiled with debug iterators disabled for
some reasons.. Even though they're actually enabled..
* Just a reminder to myself: FIX THE DAMN WAN + IMAGE EXPORT CODE !

+++ src/ext_fmts/adpcm.cpp
@@ -63,7 +63,7 @@ namespace audio
             if(index < 0)
                 return 0;
 
-            if(index >= IMA_ADPCM::StepSizes.size())
+            if( static_cast<uint32_t>(index) >= IMA_ADPCM::StepSizes.size() )
                 return (IMA_ADPCM::StepSizes.size() - 1);
 
             return index;
@@ -102,7 +102,7 @@ namespace audio
             if(index < 0)
                 return 0;
 
-            if(index >= IMA_ADPCM::StepSizes.size())
+            if( static_cast<uint32_t>(index) >= IMA_ADPCM::StepSizes.size())
                 return (IMA_ADPCM::StepSizes.size() - 1);
 
             return index;
